Address review feedback on daemon bootstrap
