+++ crates/weaverd/src/tests/mod.rs
@@ -1,5 +1,6 @@
 //! Test suites for the Weaver daemon bootstrap.
 
 mod behaviour;
+mod lib_api;
 mod support;
 mod unit;
