+++ crates/weaverd/src/tests/support/mod.rs
@@ -1,0 +1,11 @@
+//! Test harness utilities for the daemon bootstrap behavioural suite.
+
+mod backend_provider;
+mod config_loader;
+mod reporter;
+mod world;
+
+pub use backend_provider::RecordingBackendProvider;
+pub use config_loader::TestConfigLoader;
+pub use reporter::{HealthEvent, RecordingHealthReporter};
+pub use world::{TestWorld, world};
